feat: Article 삭제 시 댓글 일괄 삭제 기능 구현
- Article 삭제 시 댓글이 전부 게시글 작성자가 쓴 것이거나 댓글이 없을 경우 게시글, 댓글 삭제 가능하게 구현
- ArticleId로 is_deleted 하는 메서드 추가
- ArticleServiceTest, ReplyRepositoryTest 테스트 작성

+++ src/main/java/com/woowa/cafe/service/ArticleService.java
@@ -79,6 +79,17 @@ public class ArticleService {
             throw new HttpException(SC_NOT_FOUND, "다른 사람이 삭제할 수 없습니다.");
         }
 
+        List<Reply> replies = replyRepository.findByArticleId(articleId);
+
+        long myReplyCnt = replies.stream()
+                .filter(reply -> reply.isSameWriter(memberId))
+                .count();
+
+        if (replies.size() != myReplyCnt) {
+            throw new HttpException(SC_NOT_FOUND, "다른 사람이 작성한 댓글이 있어 삭제할 수 없습니다.");
+        }
+        replyRepository.deleteByArticleId(articleId);
+
         articleRepository.delete(articleId);
     }
 }
